refactor: streamline embedding model handling in BasicSettings and QASettings
- Removed the embedding model selection from BasicSettings, consolidating its functionality within QASettings.
- Introduced a new handler for setting the default embedding model in QASettings, ensuring user confirmation via RebuildIndexConfirmModal.
- Updated the UI components in QASettings to provide detailed descriptions and tooltips for embedding model selection, enhancing user experience and clarity.
